refactor: move pagination utilities to a dedicated module task_409
